Cygwin/X: Build fix
Commit c59dd27ecb1751f0b097046b2f892028e5a10a3e replaces __CYGWIN__ with WIN32
in a preprocessor check, as the code is appropriate for both Cygwin and Mingw.

But this isn't quite the right was to do this. Being a Win32 platform is
Cygwin's secret shame, and WIN32 isn't actually defined to avoid turning on
code specific to native-Win API ports

(See http://www.cygwin.com/faq/faq.programming.html#faq.programming.preprocessor)

Yes, this probably means WIN32 could bereplaced by __MINGW32__ everywhere in the
X sources...