Fix compilation of host tools in cross-compilation case
At 36e9f0d351afbf7fd2595990b2d39e7c551f6420, a fix was added to use
the host gcc instead of the target gcc when cross-compiling
libXt. This fix works, but is not solve the whole problem: the CFLAGS
and LDFLAGS used with the host compilers are the one defined for the
target compiler (and the flags for both compilers might be very
different).

This new fix let libXt obey to CFLAGS_FOR_BUILD and LDFLAGS_FOR_BUILD
environment variables, and use them to compile the host tools in
util/.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>