Incorrect cast in XtCvtIntToPixmap could cause bus error on 64-bit machines
Originally fixed by Leo Binchy for Sun Bug 4163152 to resolve bus error
on VSW5 (XTS) testing on 64-bit SPARC test run on Solaris 7.
( Testcase: XtCvtIntToPixmap, Purpose:  1)

Since we're converting from Int to Pixmap, need to access the argument
passed in as an int, not a Pixmap, which is consistent with the other
IntToXXX converters in this file.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Reviewed-by: Matt Turner <mattst88@gmail.com>