man: replace hard coded section number 3Xt with __libmansuffix__ #35050
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=512130

The Xt manpages are installed in section 3 (and not 3Xt as upstream
would have it). But, internal cross-references (in particular, in the
SEE ALSO part) are left alone, and keep pointing at Foo(3Xt).  The
result is that they can't be automatically followed from manpage
sanitizers such as dwww or emacs.

Reviewed-by: Alan Coopersmith <alan.coopersmith@oracle.com>
Signed-off-by: Gaetan Nadon <memsize@videotron.ca>