Fix cross-compilation
When cross-compiling against a sysroot for instance, a --sysroot option
is typically passed via the CPPFLAGS variable. This can cause the build
to break in the util subdirectory which requires the CPPFLAGS to be set
for the build system.

Solve this by using the value of CPPFLAGS_FOR_BUILD for CPPFLAGS in the
util subdirectory. Cross-compile setups can override the variable when
calling the configure script.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>
Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>