Referring to the xterm FAQ
	https://invisible-island.net/xterm/xterm.faq.html#building_it

I would like to add an optional feature which makes the String typedef
"const" if _CONST_X_STRING is defined.  Modifying the Intrinsic.h header
is only a partial solution because the library has not been modified to
take advantage of this (if it had, most of the const-warnings from gcc
would go away).

This change modifies the header file as well as the makestrs.c program
which generates Shell.h and StringDefs.h so that they use String-casts
rather than char*-casts (reducing the warnings seen when compiling with
Xt's header files).

Signed-off-by: Thomas E. Dickey <dickey@invisible-island.net>