Proposed revision of XtAppMainLoop() in
	https://gitlab.freedesktop.org/xorg/lib/libxt/issues/7

caused applications such as xclock to hang, because it did first not
check if there was an available event corresponding to the mask
parameter before calling XtAppProcessEvent().

For instance, if the mask was XtIMXEvent, it would ignore timer events
until an X event occurred.

Fix this by checking with XtAppPending() when the mask is not XtIMAll.

Signed-off-by: Thomas E. Dickey <dickey@invisible-island.net>